requirements.txt
click~=8.1.8
pydantic~=2.10.5
adafruit-circuitpython-neopixel~=6.3.14
adafruit-blinka~=8.51.0
